Made things wibble and wobble.
Mostly I just added a level to keep the list of entities,
and to have an object to query for all entities in a
certain area. This might need to be optimized eventually,
but for now it's fine.

I also had to change coordinates to doubles so I can
easily work with velocities in pixels per second. I'm
calculating this based on nano second time deltas between
main loop iterations. Note that this will make lags
very fun.