Made zombies chase humans.
Currently this is very simple. If two entities are close
enough, they check if one is a zombie and the other isn't.
If one is a zombie and the other is a human, the zombie
starts running at the human and the human starts running
away from the zombie for some time.

This will need a bunch of refactoring, obviously. I plan
to move the decision logic into an AI class so I can keep
state in there, have humans wander around, have zombies
shamble around and so on.